Moved server to clickhouse instance in test.
